# src/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Travellers API",
    "description": "API для платформи мандрівників: автентифікація, користувачі з власними статтями та можливістю зберігати улюблені історії. Є стрічка історій з пагінацією, фільтрацією за категоріями, а також завантаження фото на Cloudinary. Усі приватні ендпоінти вимагають Bearer access token.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://travellers-node.onrender.com"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/stories": {
      "get": {
        "operationId": "getStories",
        "tags": ["Stories"],
        "summary": "Отримати історії",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список історій"
          }
        }
      },
      "post": {
        "operationId": "createStory",
        "tags": ["Stories"],
        "summary": "Створити нову історію",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["title", "article", "category", "img"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "article": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Створено"
          }
        }
      }
    },
    "/api/stories/{storiesId}": {
      "patch": {
        "operationId": "updateStory",
        "tags": ["Stories"],
        "summary": "Оновити історію",
        "parameters": [
          {
            "in": "path",
            "name": "storiesId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "article": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Оновлено"
          }
        }
      }
    },
    "/api/stories/popular": {
      "get": {
        "operationId": "getPopularStories",
        "tags": ["Stories"],
        "summary": "Отримати популярні історії",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Популярні історії успішно знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found popular stories!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "stories": {
                          "type": "array",
                          "description": "Список популярних історій",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "example": "68498236a100312bea079004"
                              },
                              "img": {
                                "type": "string",
                                "example": "https://ftp.goit.study/img/travel-blog/68498236a100312bea079004.webp"
                              },
                              "title": {
                                "type": "string",
                                "example": "Сицилія: вулкани, міста i смаки"
                              },
                              "article": {
                                "type": "string",
                                "example": "Сицилія — це поєднання історії, природи i кулінарії..."
                              },
                              "category": {
                                "type": "string",
                                "example": "68fb50c80ae91338641121f2"
                              },
                              "ownerId": {
                                "type": "string",
                                "example": "6881563901add19ee16fcffb"
                              },
                              "date": {
                                "type": "string",
                                "example": "2025-03-08"
                              },
                              "favoriteCount": {
                                "type": "integer",
                                "example": 29
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 94
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 3
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}